Bug fixes and minor refactoring
