+++ kubernetesclient/namespace_client.go
@@ -6,12 +6,12 @@ import (
 	"github.com/rancher/kubernetes-model/model"
 )
 
-const NamespacePath string = "/api/v1/namespaces/%s/namespaces"
+const NamespacePath string = "/api/v1/namespaces/"
 const NamespaceByNamePath string = "/api/v1/namespaces/%s"
 
 type NamespaceOperations interface {
 	ByName(name string) (*model.Namespace, error)
-	CreateNamespace(namespace string, resource *model.Namespace) (*model.Namespace, error)
+	CreateNamespace(resource *model.Namespace) (*model.Namespace, error)
 	ReplaceNamespace(namespace string, resource *model.Namespace) (*model.Namespace, error)
 }
 
@@ -32,10 +32,9 @@ func (c *NamespaceClient) ByName(name string) (*model.Namespace, error) {
 	return resp, err
 }
 
-func (c *NamespaceClient) CreateNamespace(namespace string, resource *model.Namespace) (*model.Namespace, error) {
+func (c *NamespaceClient) CreateNamespace(resource *model.Namespace) (*model.Namespace, error) {
 	resp := &model.Namespace{}
-	path := fmt.Sprintf(NamespacePath, namespace)
-	err := c.client.doPost(path, resource, resp)
+	err := c.client.doPost(NamespacePath, resource, resp)
 	return resp, err
 }
