+++ rancherevents/util/util.go
@@ -1,0 +1,32 @@
+package eventhandlers
+
+import (
+	revents "github.com/rancher/go-machine-service/events"
+	"github.com/rancher/go-rancher/client"
+)
+
+func NewReply(event *revents.Event) *client.Publish {
+	return &client.Publish{
+		Name:         event.ReplyTo,
+		PreviousIds:  []string{event.Id},
+		ResourceType: event.ResourceType,
+		ResourceId:   event.ResourceId,
+	}
+}
+
+func PublishReply(reply *client.Publish, apiClient *client.RancherClient) error {
+	_, err := apiClient.Publish.Create(reply)
+	return err
+}
+
+func CreateAndPublishReply(event *revents.Event, cli *client.RancherClient) error {
+	reply := NewReply(event)
+	if reply.Name == "" {
+		return nil
+	}
+	err := PublishReply(reply, cli)
+	if err != nil {
+		return err
+	}
+	return nil
+}
